Removed unnecessary number format on mass and height

+++ randomstarwarscharacter/randomstarwarscharacter-data/src/main/java/com/rizzek/randomstarwarscharacter/data/remote/RemoteStarWarsCharacter.kt
@@ -27,8 +27,8 @@ data class RemoteStarWarsCharacter(
     fun toEntity(): StarWarsCharacter {
         return StarWarsCharacter(
             name = this.name,
-            height = this.height.toInt(),
-            mass = this.mass.toInt(),
+            height = this.height,
+            mass = this.mass,
             hairColor = this.hairColor,
             skinColor = this.skinColor,
             eyeColor = this.eyeColor,
